refactor: rename Terminal/window directory to Terminal/Window
