+++ tests/Window/Window.test.cpp
@@ -1,4 +1,4 @@
-#include "Terminal/window/window.hpp"
+#include "Terminal/Window/window.hpp"
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
